pyproject.toml
[tool.poetry]
name = "ttslib"
version = "0.1.1"
description = "TTS for local usage that works for all OS's, with a simple interface"
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "*"
"ruamel.yaml" = "^0.15.74"
importlib_resources = "^1.0"
google_speech = "^1.0"
pyttsx3 = "^2.7"

[tool.poetry.dev-dependencies]
